Исправлено

+++ main-service/src/main/java/ru/practicum/ewm/service/UserServiceImpl.java
@@ -6,6 +6,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.practicum.ewm.dto.NewUserRequest;
 import ru.practicum.ewm.dto.UserDto;
+import ru.practicum.ewm.exception.ConflictException;
+import ru.practicum.ewm.exception.NotFoundException;
 import ru.practicum.ewm.mapper.UserMapper;
 import ru.practicum.ewm.model.User;
 import ru.practicum.ewm.repository.UserRepository;
@@ -22,6 +24,10 @@ public class UserServiceImpl implements UserService {
     @Override
     @Transactional
     public UserDto createUser(NewUserRequest newUserRequest) {
+        if (userRepository.existsByEmail(newUserRequest.getEmail())) {
+            throw new ConflictException("User with email " + newUserRequest.getEmail() + " already exists");
+        }
+
         User user = UserMapper.toUser(newUserRequest);
         User savedUser = userRepository.save(user);
         return UserMapper.toUserDto(savedUser);
@@ -29,7 +35,9 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public List<UserDto> getUsers(List<Long> ids, Integer from, Integer size) {
-        PageRequest page = PageRequest.of(from / size, size);
+        int pageNumber = from == null || from == 0 ? 0 : from / size;
+        int pageSize = (size == null) ? 10 : (size <= 0 ? 10 : size);
+        PageRequest page = PageRequest.of(pageNumber, pageSize);
 
         List<User> users;
         if (ids != null && !ids.isEmpty()) {
@@ -46,6 +54,9 @@ public class UserServiceImpl implements UserService {
     @Override
     @Transactional
     public void deleteUser(Long userId) {
+        if (!userRepository.existsById(userId)) {
+            throw new NotFoundException("User with id=" + userId + " not found");
+        }
         userRepository.deleteById(userId);
     }
 }
